app/requirements.txt
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
psycopg2==2.9.1
PyYAML==6.0